moved shared functionality to dedicated directory
